Feature/spring security

+++ rest-service/src/main/java/com/gausman/dokolist/restservice/controller/DokoSessionController.java
@@ -3,7 +3,10 @@ package com.gausman.dokolist.restservice.controller;
 import com.gausman.dokolist.restservice.dto.CreateDokoSessionRequest;
 import com.gausman.dokolist.restservice.exception.SessionNotFoundException;
 import com.gausman.dokolist.restservice.model.entities.DokoSession;
+import com.gausman.dokolist.restservice.service.DokoGroupService;
 import com.gausman.dokolist.restservice.service.DokoSessionService;
+import com.gausman.dokolist.restservice.service.JwtService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,11 +14,20 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+
 @RestController
 @RequestMapping("/groups")
 public class DokoSessionController {
     @Autowired
     private DokoSessionService dokoSessionService;
+
+    @Autowired
+    private JwtService jwtService;
+
+    @Autowired
+    private DokoGroupService dokoGroupService;
 
     @GetMapping("/sessions/{id}")
     public ResponseEntity<DokoSession> findById(@PathVariable Long id){
@@ -47,13 +59,45 @@ public class DokoSessionController {
     }
 
     @PostMapping("/sessions/create")
-    public ResponseEntity<DokoSession> create(@RequestBody CreateDokoSessionRequest request){
+    public ResponseEntity<DokoSession> create(
+            @RequestBody CreateDokoSessionRequest request,
+            HttpServletRequest httpServletRequest
+    ){
+        String authHeader = httpServletRequest.getHeader(AUTHORIZATION);
+        String token = authHeader.substring(7);
+        String userEmail = jwtService.extractUsername(token);
+
+        // Auth check
+        if (!dokoGroupService.isPlayerInGroupByEmail(userEmail, request.getGroupId())){
+            return ResponseEntity
+                    .status(FORBIDDEN)
+                    .build();
+
+        }
+
         DokoSession session = dokoSessionService.createSession(request);
         return ResponseEntity.status(HttpStatus.CREATED).body(session);
     }
+
     @DeleteMapping("/sessions/{id}/delete")
-    public ResponseEntity<String> deleteSession(@PathVariable Long id) {
+    public ResponseEntity<String> deleteSession(
+            @PathVariable Long id,
+            HttpServletRequest httpServletRequest
+    ) {
         try {
+            String authHeader = httpServletRequest.getHeader(AUTHORIZATION);
+            String token = authHeader.substring(7);
+            String userEmail = jwtService.extractUsername(token);
+
+            DokoSession dokoSession = dokoSessionService.findById(id);
+
+            // Auth check
+            if (!dokoGroupService.isPlayerInGroupByEmail(userEmail, dokoSession.getDokoGroup().getId())){
+                return ResponseEntity
+                        .status(FORBIDDEN)
+                        .body("Spieler ist nicht teil der Gruppe");
+            }
+
             dokoSessionService.deleteSessionById(id);
             return ResponseEntity.ok("Session deleted successfully.");
         } catch (SessionNotFoundException e) {
